chore(wrangler): Move misplaced tests to dedicated Pages test folder
All tests related to Pages commands should live under
the Pages dedicated `src/__tests__/pages` folder.

This commit moves a couple of "lost" Pages tests to
that folder.
